Add files via upload
This project deals with how to create a Convolutional Neural Network (CNN) in Keras with a TensorFlow backend and to train CNNs to solve Image Classification problems. This project, creates and trains  a CNN model on a subset of the popular CIFAR-10 dataset.
Prerequisites:
In order to be successful in this project, you should be familiar with python and convolutional neural networks. 
Learn step-by-step
1) Import Libraries
2)Preprocess Data
3)Visualize Examples
4)Create Model
5)Train the Model
6)Final Predictions